moved dashboard to routes
