+++ src/app/routes/onboarding/index.tsx
@@ -1,24 +1,28 @@
-import { type FC, useEffect, useState } from 'react';
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+import { $path } from 'remix-routes';
 
 import { useTelegram } from '@/common/providers';
 import { MNEMONIC_STORE } from '@/common/utils';
 import { getCloudStorageItem } from '@/common/wallet';
 import { LoadingScreen } from '@/components';
-import { OnboardingStartPage, RestoreWalletPage } from '@/screens/onboarding';
 
 const delay = (ttl: number) => new Promise(resolve => setTimeout(resolve, ttl));
 
-const Page: FC = () => {
+const Page = () => {
   const { webApp } = useTelegram();
-  const [isLoading, setIsLoading] = useState(true);
-  const [mnemonic, setMnenonic] = useState<string | null>(null);
+  const navigate = useNavigate();
 
   useEffect(() => {
     let mounted = true;
     Promise.all([getCloudStorageItem(MNEMONIC_STORE), delay(1000)]).then(([value]) => {
-      if (mounted) {
-        setMnenonic(value ?? null);
-        setIsLoading(false);
+      if (!mounted) return;
+
+      if (value) {
+        navigate($path('/onboarding/restore/:mnemonic', { mnemonic: value }), { replace: true });
+      } else {
+        navigate($path('/onboarding/start'), { replace: true });
       }
     });
 
@@ -27,11 +31,7 @@ const Page: FC = () => {
     };
   }, [webApp]);
 
-  if (isLoading) {
-    return <LoadingScreen />;
-  }
-
-  return mnemonic ? <RestoreWalletPage mnemonic={mnemonic} /> : <OnboardingStartPage />;
+  return <LoadingScreen />;
 };
 
 export default Page;
