+++ src/components/Assets/Balance.tsx
@@ -13,6 +13,7 @@ type Props = {
 const Balance = ({ balance, precision, animate }: Props) => {
   // to prevent CountUp loosing value bug
   const [key, setKey] = useState(0);
+
   useEffect(() => setKey(x => x + 1), []);
 
   if (balance === undefined) {
