--- src/screens/onboarding/index.ts
@@ -1,2 +1,0 @@
-export { OnboardingStartPage } from './OnboardingStartPage';
-export { RestoreWalletPage } from './RestoreWallet';
