+++ vite.config.ts
@@ -2,13 +2,15 @@
 
 import { vitePlugin as remix } from '@remix-run/dev';
 import { createRoutesFromFolders } from '@remix-run/v1-route-convention';
-
 import react from '@vitejs/plugin-react';
 import { remixRoutes } from 'remix-routes/vite';
 import { defineConfig } from 'vite';
 import { compression } from 'vite-plugin-compression2';
 import svgr from 'vite-plugin-svgr';
 import tsconfigPaths from 'vite-tsconfig-paths';
+
+const isTest = (mode: string): boolean => mode === 'test';
+const isProd = (mode: string): boolean => mode === 'production';
 
 export default defineConfig(({ mode }) => ({
   envPrefix: 'PUBLIC_',
@@ -21,12 +23,12 @@ export default defineConfig(({ mode }) => ({
       },
     }),
     // remix v2 doesn't disable hmr in test mode, so we simply replace it with react plugin
-    mode === 'test'
+    isTest(mode)
       ? react()
-      : remix({ appDirectory: './src/app', routes: x => createRoutesFromFolders(x, { appDirectory: './src/app' }) }),
+      : remix({ appDirectory: './src/app', routes: r => createRoutesFromFolders(r, { appDirectory: './src/app' }) }),
     remixRoutes({ strict: true }),
-    mode === 'production' && compression({ algorithm: 'gzip', compressionOptions: { level: 9 } }),
-    mode === 'production' && compression({ algorithm: 'brotliCompress' }),
+    isProd(mode) && compression({ algorithm: 'gzip', compressionOptions: { level: 9 } }),
+    isProd(mode) && compression({ algorithm: 'brotliCompress' }),
   ],
   test: {
     globals: true,
